pyproject.toml
[project]
name = "versa"
version = "0.1.0"
description = "Versa is a Discord user app (bot) aimed to provide helpful utility commands"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "py-cord>=2.7.0",
    "python-dotenv>=1.2.1",
]
